moving the print_dt function to a new file called export_to_graphviz
